requirements.txt
Flask==0.10.1
line-bot-sdk==1.2.0
python-dotenv==0.6.4
IPython==6.1.0
autopep8==1.3.2
